requirements.txt
pandas==2.2.3
numpy==2.1.2
streamlit==1.39.0
PyPDF2==3.0.1
python-docx==1.1.2
faiss-cpu==1.9.0
sentence-transformers==3.2.1
python-dotenv==1.0.1
torch==2.5.0
ollama==0.3.3
openai==1.52.2
pydantic==2.9.2
nltk==3.9.1
tenacity==9.0.0